PR1004_improve

+++ cherBlair/week_14/PR1004.java
@@ -27,7 +27,7 @@ public class PR1004 {
         public int getDivisor(int n) {
             ArrayList<Integer> list = new ArrayList<>();
 
-            for (int i = 1; i <= (int) n / 2; i++) {
+            for (int i = 1; i <= n / 2; i++) {
                 if (n % i == 0) {
                     list.add(i);
                 }
